Optimize the code of benchmark.py
