+++ llm_bench/python/benchmark.py
@@ -4,707 +4,20 @@
 import os
 import sys
 import argparse
-import time
-from pathlib import Path
 import logging as log
-import llm_bench_utils.ov_utils
-import llm_bench_utils.pt_utils
 import llm_bench_utils.model_utils
-import torch
-import numpy as np
 from openvino.runtime import get_version
-import PIL
-import hashlib
-import llm_bench_utils.metrics_print
-import llm_bench_utils.output_csv
 import traceback
-from transformers import set_seed
-from PIL import Image
 from llm_bench_utils.memory_profile import MemConsumption
-from llm_bench_utils.hook_forward import StableDiffusionHook
+import llm_bench_utils.output_csv
 import llm_bench_utils.output_json
-import llm_bench_utils.output_file
+import llm_bench_utils.bench_text as bench_text
+import llm_bench_utils.bench_image as bench_image
+import llm_bench_utils.bench_ldm_sr as bench_ldm_sr
+import llm_bench_utils.bench_classification as bench_classification
 
-FW_UTILS = {'pt': llm_bench_utils.pt_utils, 'ov': llm_bench_utils.ov_utils}
-
-DEFAULT_INFERENCE_STEPS = 20
-LCM_DEFAULT_INFERENCE_STEPS = 4
-DEFAULT_IMAGE_WIDTH = 512
-DEFAULT_IMAGE_HEIGHT = 512
-DEFAULT_SUPER_RESOLUTION_STEPS = 50
-DEFAULT_SUPER_RESOLUTION_WIDTH = 128
-DEFAULT_SUPER_RESOLUTION_HEIGHT = 128
-DEFAULT_OUTPUT_TOKEN_SIZE = 512
 
 mem_consumption = MemConsumption()
-stable_diffusion_hook = StableDiffusionHook()
-
-
-def gen_iterate_data(
-    iter_idx='',
-    in_size='',
-    infer_count='',
-    out_size='',
-    gen_time='',
-    latency='',
-    res_md5='',
-    max_rss_mem='',
-    max_shared_mem='',
-    max_uss_mem='',
-    prompt_idx='',
-    tokenization_time=[],
-):
-    iter_data = {}
-    iter_data['iteration'] = iter_idx
-    iter_data['input_size'] = in_size
-    iter_data['infer_count'] = infer_count
-    iter_data['output_size'] = out_size
-    iter_data['generation_time'] = gen_time
-    iter_data['latency'] = latency
-    iter_data['result_md5'] = res_md5
-    iter_data['first_token_latency'] = ''
-    iter_data['other_tokens_avg_latency'] = ''
-    iter_data['first_token_infer_latency'] = ''
-    iter_data['other_tokens_infer_avg_latency'] = ''
-    iter_data['max_rss_mem_consumption'] = max_rss_mem
-    iter_data['max_shared_mem_consumption'] = max_shared_mem
-    iter_data['max_uss_mem_consumption'] = max_uss_mem
-    iter_data['prompt_idx'] = prompt_idx
-    iter_data['tokenization_time'] = tokenization_time[0] if len(tokenization_time) > 0 else ''
-    iter_data['detokenization_time'] = tokenization_time[1] if len(tokenization_time) > 1 else ''
-    return iter_data
-
-
-def run_text_generation(input_text, num, model, tokenizer, args, iter_data_list, md5_list, prompt_index, bench_hook, model_precision, proc_id):
-    set_seed(args['seed'])
-    input_text_list = [input_text] * args['batch_size']
-    if args["output_dir"] is not None and num == 0:
-        for bs_index, in_text in enumerate(input_text_list):
-            llm_bench_utils.output_file.output_input_text(in_text, args, model_precision, prompt_index, bs_index, proc_id)
-    tok_encode_start = time.perf_counter()
-    input_data = tokenizer(input_text_list, return_tensors='pt')
-    tok_encode_end = time.perf_counter()
-    tok_encode_time = (tok_encode_end - tok_encode_start) * 1000
-    input_data.pop('token_type_ids', None)
-    # Remove `token_type_ids` from inputs
-    input_tokens = input_data['input_ids'] if 'input_ids' in input_data else input_data
-    input_token_size = input_tokens[0].numel()
-    if args['batch_size'] > 1:
-        out_str = '[warm-up]' if num == 0 else '[{}]'.format(num)
-        out_str += " Batch_size={}, ".format(args['batch_size'])
-        out_str += 'all input token size after padding: {} * {}, '.format(input_token_size, args['batch_size'])
-        if args['infer_count'] is not None:
-            out_str += 'all max_output_token_size: {} * {}'.format(args['infer_count'], args['batch_size'])
-        log.info(out_str)
-
-    max_rss_mem_consumption = ''
-    max_uss_mem_consumption = ''
-    max_shared_mem_consumption = ''
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.start_collect_memory_consumption()
-    max_gen_tokens = DEFAULT_OUTPUT_TOKEN_SIZE if args['infer_count'] is None else args['infer_count']
-    start = time.perf_counter()
-    if args['infer_count'] is not None and args['end_token_stopping'] is False:
-        model.generation_config.eos_token_id = None
-        model.config.eos_token_id = None
-        result = model.generate(
-            **input_data,
-            max_new_tokens=int(max_gen_tokens),
-            num_beams=args['num_beams'],
-            use_cache=True,
-            eos_token_id=None,
-            do_sample=False
-        )
-    else:
-        result = model.generate(
-            **input_data,
-            max_new_tokens=int(max_gen_tokens),
-            num_beams=args['num_beams'],
-            use_cache=True,
-            do_sample=False
-        )
-    end = time.perf_counter()
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.end_collect_momory_consumption()
-        max_rss_mem_consumption, max_shared_mem_consumption, max_uss_mem_consumption = mem_consumption.get_max_memory_consumption()
-        mem_consumption.clear_max_memory_consumption()
-
-    generation_time = end - start
-    tok_decode_start = time.perf_counter()
-    generated_text = tokenizer.batch_decode(result)
-    tok_decode_end = time.perf_counter()
-    tok_decode_time = (tok_decode_end - tok_decode_start) * 1000
-    # Only text_gen need to minus length of input_data, because generated_text may include input_text
-    num_tokens = 0
-    result_md5_list = []
-    for bs_idx in range(args['batch_size']):
-        if 'sum' not in args['model_name'] and result[bs_idx][:input_token_size].equal(input_tokens[bs_idx]):
-            generated_token_size = len(result[bs_idx]) - input_tokens[bs_idx].numel()
-        else:
-            generated_token_size = len(result[bs_idx])
-        # Encoder-decoder models expect the `decoder_input_ids` to start with a special token
-        # When counting the output length, subtract 1. The last token does not participate in inference.
-        if model.config.is_encoder_decoder and result[bs_idx][0] == model.config.decoder_start_token_id:
-            generated_token_size = generated_token_size - 1
-        num_tokens += generated_token_size
-        if generated_token_size > max_gen_tokens:
-            log.error('Output token size is over max output token size!')
-        result_text = generated_text[bs_idx]
-        if args["output_dir"] is not None:
-            llm_bench_utils.output_file.output_gen_text(result_text, args, model_precision, prompt_index, num, bs_idx, proc_id)
-        result_md5_list.append(hashlib.new("md5", result_text.encode(), usedforsecurity=False).hexdigest())
-    if len(md5_list[num]) == 0:
-        md5_list[num] = {prompt_index : result_md5_list}
-    else:
-        md5_list[num][prompt_index] = result_md5_list
-    per_token_time = generation_time * 1000 / (num_tokens / args['batch_size'])
-    tm_list = []
-    tm_infer_list = []
-    if bench_hook is not None:
-        tm_list = bench_hook.get_time_list()
-        log.debug('latency of all tokens:')
-        [log.debug('[{}]{:.4f}'.format(idx, tm)) for idx, tm in enumerate(tm_list)]
-        tm_infer_list = bench_hook.get_time_infer_list()
-        log.debug('latency of all infers:')
-        [log.debug('[{}]{:.4f}'.format(idx, tm)) for idx, tm in enumerate(tm_infer_list)]
-        if args['num_beams'] == 1 and generated_token_size != len(tm_infer_list):
-            log.warning(f'Output token size({generated_token_size}) is not equal to infer count({len(tm_infer_list)})')
-    iter_data = gen_iterate_data(
-        num,
-        input_token_size * args['batch_size'],
-        len(tm_infer_list),
-        num_tokens,
-        generation_time,
-        per_token_time,
-        result_md5_list,
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        prompt_idx=prompt_index,
-        tokenization_time=(tok_encode_time, tok_decode_time)
-    )
-    iter_data_list.append(iter_data)
-    llm_bench_utils.metrics_print.print_metrics(
-        num,
-        iter_data,
-        tm_list,
-        tm_infer_list,
-        warm_up=(num == 0),
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        tokenization_time=(tok_encode_time, tok_decode_time),
-        batch_size=args['batch_size']
-    )
-    if num > 0:
-        prev_md5 = md5_list[num - 1][prompt_index]
-        if result_md5_list != prev_md5:
-            log.warning(f"[{num}] Prompt[{prompt_index}]'s md5 {result_md5_list} "
-                        f"is different from md5 of the {num - 1} iteration {prev_md5}")
-            llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-            if not args.get("use_cb", False):
-                if num == 1:
-                    # if the device is CPU, throw exception
-                    if args['devices'].lower().startswith('cpu') is True:
-                        assert (result_md5_list == prev_md5)
-                else:
-                    # throw exception
-                    assert (result_md5_list == prev_md5)
-    else:
-        llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-    if bench_hook is not None:
-        bench_hook.clear_time_list()
-        bench_hook.clear_time_infer_list()
-
-
-def run_text_generation_genai(input_text, num, model, tokenizer, args, iter_data_list, md5_list, prompt_index, streamer, model_precision, proc_id):
-    set_seed(args['seed'])
-    input_text_list = [input_text] * args['batch_size']
-    if args["output_dir"] is not None and num == 0:
-        for bs_index, in_text in enumerate(input_text_list):
-            llm_bench_utils.output_file.output_input_text(in_text, args, model_precision, prompt_index, bs_index, proc_id)
-    pt_inputs = tokenizer(input_text_list, return_tensors="pt")
-    input_token_size = pt_inputs.input_ids.shape[1]
-    if args['batch_size'] > 1:
-        out_str = '[warm-up]' if num == 0 else '[{}]'.format(num)
-        out_str += " Batch_size={}, ".format(args['batch_size'])
-        out_str += 'all input token size after padding: {} * {}, '.format(input_token_size, args['batch_size'])
-        if args['infer_count'] is not None:
-            out_str += 'all max_output_token_size: {} * {}'.format(args['infer_count'], args['batch_size'])
-        log.info(out_str)
-
-    max_rss_mem_consumption = ''
-    max_uss_mem_consumption = ''
-    max_shared_mem_consumption = ''
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.start_collect_memory_consumption()
-    max_gen_tokens = DEFAULT_OUTPUT_TOKEN_SIZE if args['infer_count'] is None else args['infer_count']
-    start = time.perf_counter()
-    generation_result = model.generate(input_text_list, max_new_tokens=max_gen_tokens, num_beams=args["num_beams"])
-    end = time.perf_counter()
-    generated_text = generation_result.texts
-    perf_metrics = generation_result.perf_metrics
-
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.end_collect_momory_consumption()
-        max_rss_mem_consumption, max_shared_mem_consumption, max_uss_mem_consumption = mem_consumption.get_max_memory_consumption()
-        mem_consumption.clear_max_memory_consumption()
-
-    generation_time = end - start
-    generated_tokens = [tokenizer(text).input_ids for text in generated_text]
-    # Only text_gen need to minus length of input_data, because generated_text may include input_text
-    num_tokens = 0
-    result_md5_list = []
-    for bs_idx in range(args['batch_size']):
-        generated_text_len = len(generated_tokens[bs_idx])
-        num_tokens += generated_text_len
-        if generated_text_len > max_gen_tokens:
-            log.error('Output token size is over max output token size!')
-        result_text = generated_text[bs_idx]
-        if args["output_dir"] is not None:
-            llm_bench_utils.output_file.output_gen_text(result_text, args, model_precision, prompt_index, num, bs_idx, proc_id)
-        result_md5_list.append(hashlib.new("md5", result_text.encode(), usedforsecurity=False).hexdigest())
-    if len(md5_list[num]) == 0:
-        md5_list[num] = {prompt_index : result_md5_list}
-    else:
-        md5_list[num][prompt_index] = result_md5_list
-    per_token_time = generation_time * 1000 / (num_tokens / args['batch_size'])
-    tm_list = np.array(perf_metrics.raw_metrics.m_durations) / 1000 / 1000
-    log.debug('latency of all tokens:')
-    [log.debug('[{}]{:.4f}'.format(idx, tm)) for idx, tm in enumerate(tm_list)]
-    tokenization_time = (
-        np.mean(perf_metrics.raw_metrics.tokenization_durations) / 1000,
-        np.mean(perf_metrics.raw_metrics.detokenization_durations) / 1000
-    )
-    iter_data = gen_iterate_data(
-        num,
-        input_token_size * args['batch_size'],
-        len(tm_list),
-        num_tokens,
-        generation_time,
-        per_token_time,
-        result_md5_list,
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        prompt_idx=prompt_index,
-        tokenization_time=tokenization_time
-    )
-    iter_data_list.append(iter_data)
-    llm_bench_utils.metrics_print.print_metrics(
-        num,
-        iter_data,
-        tm_list.tolist(),
-        [],
-        warm_up=(num == 0),
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        tokenization_time=tokenization_time,
-        batch_size=args['batch_size']
-    )
-    if num > 0:
-        prev_md5 = md5_list[num - 1][prompt_index]
-        if result_md5_list != prev_md5:
-            log.warning(f"[{num}] Prompt[{prompt_index}]'s md5 {result_md5_list} "
-                        f"is different from md5 of the {num - 1} iteration {prev_md5}")
-            llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-            if not args.get("use_cb", False):
-                if num == 1:
-                    # if the device is CPU, throw exception
-                    if args['devices'].lower().startswith('cpu') is True:
-                        assert (result_md5_list == prev_md5)
-                else:
-                    # throw exception
-                    assert (result_md5_list == prev_md5)
-    else:
-        llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-
-
-def run_text_generation_genai_with_stream(input_text, num, model, tokenizer, args, iter_data_list, md5_list, prompt_index, streamer, model_precision, proc_id):
-    set_seed(args['seed'])
-    input_text_list = [input_text] * args['batch_size']
-    if args["output_dir"] is not None and num == 0:
-        for bs_index, in_text in enumerate(input_text_list):
-            llm_bench_utils.output_file.output_input_text(in_text, args, model_precision, prompt_index, bs_index, proc_id)
-    pt_inputs = tokenizer(input_text_list, return_tensors="pt")
-    input_token_size = pt_inputs.input_ids.shape[1]
-    pipe_tokenizer = model.get_tokenizer()
-    tok_encode_start = time.perf_counter()
-    input_data = pipe_tokenizer.encode(input_text_list)
-    tok_encode_end = time.perf_counter()
-    tok_encode_time = (tok_encode_end - tok_encode_start) * 1000
-    if args['batch_size'] > 1:
-        out_str = '[warm-up]' if num == 0 else '[{}]'.format(num)
-        out_str += " Batch_size={}, ".format(args['batch_size'])
-        out_str += 'all input token size after padding: {} * {}, '.format(input_token_size, args['batch_size'])
-        if args['infer_count'] is not None:
-            out_str += 'all max_output_token_size: {} * {}'.format(args['infer_count'], args['batch_size'])
-        log.info(out_str)
-    max_rss_mem_consumption = ''
-    max_uss_mem_consumption = ''
-    max_shared_mem_consumption = ''
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.start_collect_memory_consumption()
-    max_gen_tokens = DEFAULT_OUTPUT_TOKEN_SIZE if args['infer_count'] is None else args['infer_count']
-    streamer.reset()
-    start = time.perf_counter()
-    generated_tokens = model.generate(input_data, max_new_tokens=max_gen_tokens, num_beams=args["num_beams"], streamer=streamer).tokens
-    end = time.perf_counter()
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.end_collect_momory_consumption()
-        max_rss_mem_consumption, max_shared_mem_consumption, max_uss_mem_consumption = mem_consumption.get_max_memory_consumption()
-        mem_consumption.clear_max_memory_consumption()
-    generation_time = end - start
-    tok_decode_start = time.perf_counter()
-    generated_text = pipe_tokenizer.decode(generated_tokens)
-    tok_decode_end = time.perf_counter()
-    tok_decode_time = (tok_decode_end - tok_decode_start) * 1000
-    # Only text_gen need to minus length of input_data, because generated_text may include input_text
-    num_tokens = 0
-    result_md5_list = []
-    for bs_idx in range(args['batch_size']):
-        generated_text_len = len(generated_tokens[bs_idx])
-        num_tokens += generated_text_len
-        if generated_text_len > max_gen_tokens:
-            log.error('Output token size is over max output token size!')
-        result_text = generated_text[bs_idx]
-        if args["output_dir"] is not None:
-            llm_bench_utils.output_file.output_gen_text(result_text, args, model_precision, prompt_index, num, bs_idx, proc_id)
-        result_md5_list.append(hashlib.new("md5", result_text.encode(), usedforsecurity=False).hexdigest())
-    if len(md5_list[num]) == 0:
-        md5_list[num] = {prompt_index : result_md5_list}
-    else:
-        md5_list[num][prompt_index] = result_md5_list
-    per_token_time = generation_time * 1000 / (num_tokens / args['batch_size'])
-    tm_list = streamer.get_time_list()
-    log.debug('latency of all tokens:')
-    [log.debug('[{}]{:.4f}'.format(idx, tm)) for idx, tm in enumerate(tm_list)]
-    iter_data = gen_iterate_data(
-        num,
-        input_token_size * args['batch_size'],
-        len(tm_list),
-        num_tokens,
-        generation_time,
-        per_token_time,
-        result_md5_list,
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        prompt_idx=prompt_index,
-        tokenization_time=(tok_encode_time, tok_decode_time)
-    )
-    iter_data_list.append(iter_data)
-    llm_bench_utils.metrics_print.print_metrics(
-        num,
-        iter_data,
-        tm_list,
-        [],
-        warm_up=(num == 0),
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        tokenization_time=(tok_encode_time, tok_decode_time),
-        batch_size=args['batch_size']
-    )
-    if num > 0:
-        prev_md5 = md5_list[num - 1][prompt_index]
-        if result_md5_list != prev_md5:
-            log.warning(f"[{num}] Prompt[{prompt_index}]'s md5 {result_md5_list} "
-                        f"is different from md5 of the {num - 1} iteration {prev_md5}")
-            llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-            if not args.get("use_cb", False):
-                if num == 1:
-                    # if the device is CPU, throw exception
-                    if args['devices'].lower().startswith('cpu') is True:
-                        assert (result_md5_list == prev_md5)
-                else:
-                    # throw exception
-                    assert (result_md5_list == prev_md5)
-    else:
-        llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=generated_text[0])
-    streamer.reset()
-
-
-def run_text_generation_benchmark(model_path, framework, device, args, num_iters):
-    model, tokenizer, pretrain_time, bench_hook, use_genai = FW_UTILS[framework].create_text_gen_model(model_path, device, **args)
-    model_precision = llm_bench_utils.model_utils.get_model_precision(model_path.parts)
-    iter_data_list = []
-    md5_list = {num : {} for num in range(num_iters + 1)}
-    input_text_list = llm_bench_utils.model_utils.get_prompts(args)
-    if args['prompt_index'] is None:
-        prompt_idx_list = [prompt_idx for prompt_idx, input_text in enumerate(input_text_list)]
-        text_list = input_text_list
-    else:
-        prompt_idx_list = []
-        text_list = []
-        for i in args['prompt_index']:
-            if 0 <= i < len(input_text_list):
-                text_list.append(input_text_list[i])
-                prompt_idx_list.append(i)
-    if len(input_text_list) == 0:
-        raise RuntimeError('==Failure prompts is empty ==')
-    log.info(f"Numbeams: {args['num_beams']}, benchmarking iter nums(exclude warm-up): {num_iters}, "
-             f'prompt nums: {len(text_list)}, prompt idx: {prompt_idx_list}')
-
-    # if num_iters == 0, just output warm-up data
-    if not use_genai:
-        text_gen_fn = run_text_generation
-    elif bench_hook is not None:
-        text_gen_fn = run_text_generation_genai_with_stream
-    else:
-        text_gen_fn = run_text_generation_genai
-    proc_id = os.getpid()
-    if args['subsequent'] is False:
-        for num in range(num_iters + 1):
-            for idx, input_text in enumerate(text_list):
-                if num == 0:
-                    log.info(f'[warm-up] Input text: {input_text}')
-                text_gen_fn(input_text, num, model, tokenizer, args, iter_data_list, md5_list, prompt_idx_list[idx], bench_hook, model_precision, proc_id)
-    else:
-        for idx, input_text in enumerate(text_list):
-            for num in range(num_iters + 1):
-                if num == 0:
-                    log.info(f'[warm-up] Input text: {input_text}')
-                text_gen_fn(input_text, num, model, tokenizer, args, iter_data_list, md5_list, prompt_idx_list[idx], bench_hook, model_precision, proc_id)
-
-    llm_bench_utils.metrics_print.print_average(iter_data_list, prompt_idx_list, args['batch_size'], True)
-    return iter_data_list, pretrain_time
-
-
-def run_image_generation(image_param, num, image_id, pipe, args, iter_data_list, proc_id):
-    set_seed(args['seed'])
-    input_text = image_param['prompt']
-    image_width = image_param.get('width', DEFAULT_IMAGE_WIDTH)
-    image_height = image_param.get('height', DEFAULT_IMAGE_HEIGHT)
-    nsteps = image_param.get('steps', DEFAULT_INFERENCE_STEPS if 'lcm' not in args["model_name"] else LCM_DEFAULT_INFERENCE_STEPS)
-    guidance_scale = image_param.get('guidance_scale', None)
-    log.info(
-        f"[{'warm-up' if num == 0 else num}] Input params: Batch_size={args['batch_size']}, "
-        f'steps={nsteps}, width={image_width}, height={image_height}, guidance_scale={guidance_scale}'
-    )
-    result_md5_list = []
-    max_rss_mem_consumption = ''
-    max_uss_mem_consumption = ''
-    max_shared_mem_consumption = ''
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.start_collect_memory_consumption()
-    additional_args = {}
-    if guidance_scale is not None:
-        additional_args["guidance_scale"] = guidance_scale
-    else:
-        if 'lcm-sdxl' in args['model_type']:
-            additional_args["guidance_scale"] = 1.0
-        if 'turbo' in args['model_name']:
-            additional_args["guidance_scale"] = 0.0
-    input_text_list = [input_text] * args['batch_size']
-    if num == 0 and args["output_dir"] is not None:
-        for bs_idx, in_text in enumerate(input_text_list):
-            llm_bench_utils.output_file.output_image_input_text(in_text, args, image_id, bs_idx, proc_id)
-    start = time.perf_counter()
-    res = pipe(input_text_list, num_inference_steps=nsteps, height=image_height, width=image_width, **additional_args).images
-    end = time.perf_counter()
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.end_collect_momory_consumption()
-        max_rss_mem_consumption, max_shared_mem_consumption, max_uss_mem_consumption = mem_consumption.get_max_memory_consumption()
-        mem_consumption.clear_max_memory_consumption()
-    for bs_idx in range(args['batch_size']):
-        rslt_img_fn = llm_bench_utils.output_file.output_gen_image(res[bs_idx], args, image_id, num, bs_idx, proc_id, '.png')
-        result_md5_list.append(hashlib.md5(Image.open(rslt_img_fn).tobytes(), usedforsecurity=False).hexdigest())
-    generation_time = end - start
-    iter_data = gen_iterate_data(
-        iter_idx=num,
-        infer_count=nsteps,
-        gen_time=generation_time,
-        res_md5=result_md5_list,
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        prompt_idx=image_id,
-    )
-    iter_data_list.append(iter_data)
-    llm_bench_utils.metrics_print.print_metrics(
-        num,
-        iter_data,
-        warm_up=(num == 0),
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        stable_diffusion=stable_diffusion_hook
-    )
-    llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=rslt_img_fn)
-    stable_diffusion_hook.clear_statistics()
-
-
-def run_image_generation_benchmark(model_path, framework, device, args, num_iters):
-    if args['genai']:
-        log.warning("GenAI pipeline is not supported for this task. Switched on default benchmarking")
-    pipe, pretrain_time = FW_UTILS[framework].create_image_gen_model(model_path, device, **args)
-    iter_data_list = []
-    input_image_list = llm_bench_utils.model_utils.get_image_param_from_prompt_file(args)
-    if framework == "ov":
-        stable_diffusion_hook.new_text_encoder(pipe)
-        stable_diffusion_hook.new_unet(pipe)
-        stable_diffusion_hook.new_vae_decoder(pipe)
-
-    if args['prompt_index'] is None:
-        prompt_idx_list = [image_id for image_id, input_text in enumerate(input_image_list)]
-        image_list = input_image_list
-    else:
-        prompt_idx_list = []
-        image_list = []
-        for i in args['prompt_index']:
-            if 0 <= i < len(input_image_list):
-                image_list.append(input_image_list[i])
-                prompt_idx_list.append(i)
-    if len(image_list) == 0:
-        raise RuntimeError('==Failure prompts is empty ==')
-    log.info(f'Benchmarking iter nums(exclude warm-up): {num_iters}, prompt nums: {len(image_list)}, prompt idx: {prompt_idx_list}')
-
-    # if num_iters == 0, just output warm-up data
-    proc_id = os.getpid()
-    if args['subsequent'] is False:
-        for num in range(num_iters + 1):
-            for image_id, image_param in enumerate(image_list):
-                run_image_generation(image_param, num, prompt_idx_list[image_id], pipe, args, iter_data_list, proc_id)
-    else:
-        for image_id, image_param in enumerate(image_list):
-            for num in range(num_iters + 1):
-                run_image_generation(image_param, num, prompt_idx_list[image_id], pipe, args, iter_data_list, proc_id)
-
-    llm_bench_utils.metrics_print.print_average(iter_data_list, prompt_idx_list, args['batch_size'], False)
-    return iter_data_list, pretrain_time
-
-
-def run_image_classification(model_path, framework, device, args, num_iters=10):
-    if args['genai']:
-        log.warning("GenAI pipeline is not supported for this task. Switched on default benchmarking")
-    model, input_size = FW_UTILS[framework].create_image_classification_model(model_path, device, **args)
-
-    data = torch.rand(input_size)
-
-    test_time = []
-    iter_data_list = []
-    for num in range(num_iters or 10):
-        start = time.perf_counter()
-        model(data)
-        end = time.perf_counter()
-        generation_time = end - start
-        test_time.append(generation_time)
-
-        iter_data = gen_iterate_data(iter_idx=num, in_size=input_size, infer_count=num_iters, gen_time=generation_time)
-        iter_data_list.append(iter_data)
-    log.info(f'Processed {num_iters} images in {np.sum(test_time)}s')
-    log.info(f'Average processing time {np.mean(test_time)} s')
-    return iter_data_list
-
-
-def run_ldm_super_resolution(img, num, pipe, args, framework, iter_data_list, image_id, tm_list, proc_id):
-    set_seed(args['seed'])
-    nsteps = img.get('steps', DEFAULT_SUPER_RESOLUTION_STEPS)
-    resize_image_width = img.get('width', DEFAULT_SUPER_RESOLUTION_WIDTH)
-    resize_image_height = img.get('height', DEFAULT_SUPER_RESOLUTION_HEIGHT)
-    log.info(
-        f"[{'warm-up' if num == 0 else num}] Input params: steps={nsteps}, "
-        f'resize_width={resize_image_width}, resize_height={resize_image_height}'
-    )
-    low_res_img = PIL.Image.open(img['prompt']).convert('RGB')
-    low_res_img = low_res_img.resize((resize_image_width, resize_image_height))
-    max_rss_mem_consumption = ''
-    max_uss_mem_consumption = ''
-    max_shared_mem_consumption = ''
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.start_collect_memory_consumption()
-    start = time.perf_counter()
-    res = pipe(low_res_img, num_inference_steps=nsteps, tm_list=tm_list)
-    end = time.perf_counter()
-    if (args['mem_consumption'] == 1 and num == 0) or args['mem_consumption'] == 2:
-        mem_consumption.end_collect_momory_consumption()
-        max_rss_mem_consumption, max_shared_mem_consumption, max_uss_mem_consumption = mem_consumption.get_max_memory_consumption()
-        mem_consumption.clear_max_memory_consumption()
-    result_md5_list = []
-    if framework == 'ov':
-        rslt_img_fn = llm_bench_utils.output_file.output_gen_image(res[0], args, image_id, num, None, proc_id, '.png')
-        result_md5_list.append(hashlib.md5(Image.open(rslt_img_fn).tobytes(), usedforsecurity=False).hexdigest())
-
-    generation_time = end - start
-    iter_data = gen_iterate_data(
-        iter_idx=num,
-        infer_count=nsteps,
-        gen_time=generation_time,
-        res_md5=result_md5_list,
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption,
-        prompt_idx=image_id,
-    )
-    iter_data_list.append(iter_data)
-    llm_bench_utils.metrics_print.print_metrics(
-        num,
-        iter_data,
-        warm_up=(num == 0),
-        max_rss_mem=max_rss_mem_consumption,
-        max_shared_mem=max_shared_mem_consumption,
-        max_uss_mem=max_uss_mem_consumption
-    )
-    llm_bench_utils.metrics_print.print_generated(num, warm_up=(num == 0), generated=rslt_img_fn)
-    llm_bench_utils.metrics_print.print_ldm_unet_vqvae_infer_latency(num, iter_data, tm_list, warm_up=(num == 0))
-
-
-def run_ldm_super_resolution_benchmark(model_path, framework, device, args, num_iters):
-    if args["genai"]:
-        log.warning("GenAI pipeline is not supported for this task. Switched on default benchmarking")
-    pipe, pretrain_time = FW_UTILS[framework].create_ldm_super_resolution_model(model_path, device, **args)
-    iter_data_list = []
-    tm_list = []
-    input_image_list = llm_bench_utils.model_utils.get_image_param_from_prompt_file(args)
-    if len(input_image_list) > 0:
-        images = []
-        for image in input_image_list:
-            if args['prompt'] is None and args['prompt_file'] is None:
-                raise RuntimeError('==Failure image is empty ==')
-            elif args['prompt_file'] is not None and len(args['prompt_file']) > 0:
-                image['prompt'] = os.path.join(os.path.dirname(args['prompt_file'][0]), image['prompt'].replace('./', ''))
-            image['prompt'] = Path(image['prompt'])
-            images.append(image)
-    else:
-        if args['images'] is not None:
-            images = Path(args['images'])
-            if images.is_dir():
-                images = list(images.glob('*'))
-            else:
-                images = [images]
-        else:
-            raise RuntimeError('==Failure image is empty ==')
-
-    prompt_idx_list = [image_id for image_id, image_param in enumerate(images)]
-    if args['prompt_index'] is None:
-        prompt_idx_list = [image_id for image_id, input_text in enumerate(images)]
-        image_list = images
-    else:
-        prompt_idx_list = []
-        image_list = []
-        for i in args['prompt_index']:
-            if 0 <= i < len(images):
-                image_list.append(images[i])
-                prompt_idx_list.append(i)
-    if len(image_list) == 0:
-        raise RuntimeError('==Failure prompts is empty ==')
-    log.info(f'Benchmarking iter nums(exclude warm-up): {num_iters}, prompt nums: {len(image_list)}, prompt idx: {prompt_idx_list}')
-
-    # if num_iters == 0, just output warm-up data
-    proc_id = os.getpid()
-    for num in range(num_iters + 1):
-        for image_id, img in enumerate(image_list):
-            if num == 0:
-                if args["output_dir"] is not None:
-                    llm_bench_utils.output_file.output_image_input_text(str(img['prompt']), args, prompt_idx_list[image_id], None, proc_id)
-            log.info(f"[{'warm-up' if num == 0 else num}] Input image={img['prompt']}")
-            run_ldm_super_resolution(img, num, pipe, args, framework, iter_data_list, prompt_idx_list[image_id], tm_list, proc_id)
-            tm_list.clear()
-    llm_bench_utils.metrics_print.print_average(iter_data_list, prompt_idx_list, 1, False)
-
-    return iter_data_list, pretrain_time
 
 
 def num_iters_type(x):
@@ -828,11 +141,11 @@ def get_argprser():
 
 
 CASE_TO_BENCH = {
-    'text_gen': run_text_generation_benchmark,
-    'image_gen': run_image_generation_benchmark,
-    'image_cls': run_image_classification,
-    'code_gen': run_text_generation_benchmark,
-    'ldm_super_resolution': run_ldm_super_resolution_benchmark,
+    'text_gen': bench_text.run_text_generation_benchmark,
+    'image_gen': bench_image.run_image_generation_benchmark,
+    'image_cls': bench_classification.run_image_classification,
+    'code_gen': bench_text.run_text_generation_benchmark,
+    'ldm_super_resolution': bench_ldm_sr.run_ldm_super_resolution_benchmark,
 }
 
 
@@ -857,7 +170,8 @@ def main():
     if args.memory_consumption:
         mem_consumption.start_collect_mem_consumption_thread()
     try:
-        iter_data_list, pretrain_time = CASE_TO_BENCH[model_args['use_case']](model_path, framework, args.device, model_args, args.num_iters)
+        iter_data_list, pretrain_time = CASE_TO_BENCH[model_args['use_case']](
+            model_path, framework, args.device, model_args, args.num_iters, mem_consumption)
         if args.report is not None or args.report_json is not None:
             model_precision = ''
             if framework == 'ov':
